+++ app/src/components/Cards/Playlist.tsx
@@ -38,7 +38,13 @@ export default function Playlist({ playlist }: { playlist: PlaylistType }) {
               textDecoration: "none",
             }}
           >
-            <Typography component="span" style={{ lineHeight: 1 }}>
+            <Typography
+              component="span"
+              sx={{
+                lineHeight: 1,
+                ":hover": { textDecoration: "underline" },
+              }}
+            >
               <strong>{playlist.title}</strong>
             </Typography>
           </Link>