+++ app/src/components/Processing/ProcessingItem.tsx
@@ -56,6 +56,7 @@ export const ProcessingItem = ({ item }: { item: ProcessingItemType }) => {
       </TableCell>
       <TableCell scope="row">{item.artist}</TableCell>
       <TableCell scope="row">{item.type}</TableCell>
+      <TableCell scope="row">{item.quality}</TableCell>
     </TableRow>
   );
 };